Convert BRSharedPrefs to kotlin
This change is API compatible and modifies no external code.

BRSharedPrefs becomes a Kotlin `object` so it is a singleton in Kotlin
code and in fact maintains compatibility with existing Kotlin code.

This change alone would cause existing Java code to fail because
BRSharedPrefs is just a class under the hood. Kotlin silently translates
`BRSharedPrefs.getPreferredFiatIso(ctx)` into
`BRSharedPrefs.INSTANCE.getPreferredFiatIso(ctx)`.

To avoid adding `INSTANCE` into every `BRSharedPrefs` call that exists
in java code, we annotate all the object’s methods with `@JvmStatic`
this tells the compiler to make methods static on the class
when compiling for the JVM.


Former-commit-id: aa7c1f27aac813c6a6677051e0de1adef16fd395